+++ resources/release-notes.md
@@ -8,4 +8,5 @@
 # Software
 
 - Add `isPlottable` method to check if `mag.TimeSeries` has enough data to be plotted
+- Fix display of frequencies for PSD and spectrogram plots
 - Fix issue with timed events (e.g., timed Burst mode) followed up by non-mode events (e.g., range-change events)